cloud_run/adcrawler/requirements.txt
requests==2.22.0
beautifulsoup4==4.8.2
docker==4.2.0
Flask==1.0.2
gunicorn==19.9.0
google-cloud-storage==1.26.0
lxml==4.4.2
